+++ src/contexts/QuickMarcContext/QuickMarcContext.js
@@ -5,6 +5,10 @@ import {
   useRef,
   useState,
 } from 'react';
+import {
+  useLocation,
+  useHistory,
+} from 'react-router-dom';
 import PropTypes from 'prop-types';
 
 const QuickMarcContext = createContext();
@@ -25,12 +29,34 @@ const QuickMarcProvider = ({
   marcType,
   basePath,
 }) => {
+  const location = useLocation();
+  const history = useHistory();
   const [instance, setInstance] = useState(null);
   const [marcRecord, setMarcRecord] = useState(null);
   const [selectedSourceFile, setSelectedSourceFile] = useState(null);
   const [_relatedRecordVersion, setRelatedRecordVersion] = useState();
   const validationErrors = useRef({});
   const continueAfterSave = useRef(false);
+  const isSharedRef = useRef(new URLSearchParams(location.search).get('shared') === 'true');
+
+  const setIsShared = useCallback((_isShared) => {
+    const searchParams = new URLSearchParams(location.search);
+
+    if (_isShared) {
+      searchParams.append('shared', true);
+    } else {
+      searchParams.delete('shared');
+    }
+
+    // still need to apply `shared` parameter to the url and push to history
+    // because if a user refreshes the page after "Save & keep editing" - all values in context
+    // will be lost and we'll have an incorrect isShared state
+    isSharedRef.current = _isShared;
+    history.push({
+      pathname: location.pathname,
+      search: searchParams.toString(),
+    });
+  }, [location, history]);
 
   const setValidationErrors = useCallback((newValidationErrors) => {
     validationErrors.current = newValidationErrors;
@@ -51,6 +77,8 @@ const QuickMarcProvider = ({
     initialValues: marcRecord,
     setMarcRecord,
     basePath,
+    isSharedRef,
+    setIsShared,
   }), [
     selectedSourceFile,
     setSelectedSourceFile,
@@ -66,6 +94,8 @@ const QuickMarcProvider = ({
     marcRecord,
     setMarcRecord,
     basePath,
+    isSharedRef,
+    setIsShared,
   ]);
 
   return (